% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-genome.R
\name{filter_genome_by_pathway}
\alias{filter_genome_by_pathway}
\title{Filter out genomes by pathway}
\usage{
filter_genome_by_pathway(data, pathway_name)
}
\arguments{
\item{data}{a tibble that has genome_ids - likely the output from get_genome_id()}
}
\description{
Filter out genomes by pathway
}
