% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-genome.R
\name{filter_orgs}
\alias{filter_orgs}
\alias{filter_orgs.keggerator}
\title{Filter OTUs and kegg organisms}
\usage{
filter_orgs(data, uncertainty, pathway_name, pathways, verbose, progress)

\method{filter_orgs}{keggerator}(data, uncertainty = 1,
  pathway_name = NULL, pathways = NULL, verbose = FALSE,
  progress = TRUE)
}
\arguments{
\item{data}{a tibble that has genome_ids - likely the output from get_genome_id()}

\item{pathway_name}{the name of the pathway that you wish to filter}

\item{pathways}{a kegg_tbl dataset of available kegg pathways. This dataset
is produced from the get_kegg_pathway function. If it is NULL (default) the
KEGGerator::kegg_pathways dataset will be used}
}
\description{
Filter OTUs and kegg organisms
}
\section{Methods (by class)}{
\itemize{
\item \code{keggerator}: method for keggerator objects
}}

