% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble-genomes.R
\name{genomes_tibble}
\alias{genomes_tibble}
\title{extract genomes from tax table or tax tibble}
\usage{
genomes_tibble(data, drop_taxa = TRUE)
}
\arguments{
\item{data}{either a phyloseq object or the output of tibble_tax}
}
\description{
extract genomes from tax table or tax tibble
}
