% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_datasets}
\alias{get_datasets}
\alias{get_kegg_pathway}
\alias{get_kegg_module}
\alias{get_kegg_enzyme}
\alias{get_kegg_orthology}
\title{Get Updated KEGG datasets}
\usage{
get_kegg_pathway()

get_kegg_module()

get_kegg_enzyme()

get_kegg_orthology()
}
\description{
Get Updated KEGG datasets
}
\details{
These functions will allow users to get updated datasets rather if they do
not want to use the ones supplied in this package. While I will try to keep
these datasets updated, it will allow for the flexibility for users to
have the most updated data in KEGG
}
