% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-enzymes.R
\name{get_enzymes}
\alias{get_enzymes}
\alias{get_enzymes.keggerator}
\alias{get_enzymes.orgs_id}
\title{Query all enzymes related to each genome in the dataset}
\usage{
get_enzymes(data, pathway_enzymes, kegg_enzymes, verbose, progress)

\method{get_enzymes}{keggerator}(data, pathway_enzymes = NULL,
  kegg_enzymes = NULL, verbose = FALSE, progress = TRUE)

\method{get_enzymes}{orgs_id}(orgs_id, pathway_enzymes = NULL,
  kegg_enzymes = NULL, verbose = FALSE, progress = TRUE)
}
\arguments{
\item{data}{a tibble with genomes and genome_id or the output from `get_genome_id()`}

\item{pathway_enzymes}{a vector of pathway enzymes}

\item{kegg_enzymes}{a kegg_tbl that has the columns enzyme and enzyme_id. This
can be generated using the get_kegg_enzyme() function. If NULL (default) than
the KEGGerator::kegg_enzymes dataset will be used}

\item{verbose}{logical; if TRUE the number of enzyme links for each genome
id will be shown as they are processed}

\item{progress}{logical; if TRUE (default) than a progress bar will appear if
the query takes longer than 10 seconds (it likely will if your data has more)
than only a few genomes}
}
\description{
Query all enzymes related to each genome in the dataset
}
\details{
if `pathway_enzymes` is NULL (default) than all enzymes for each genome
will be returned. If a vector of enzyme_ids is provided than only the enzymes related
to each genome that are in this vector will be returned. This is useful if used in
coordination with the output of `get_pathway_enzymes` to only include the enzymes
that are related with the pathway of interest.
}
\section{Methods (by class)}{
\itemize{
\item \code{keggerator}: method for keggerator objects

\item \code{orgs_id}: method for orgs_id objects
}}

