% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-genome-enzymes.R
\name{get_genome_enzymes}
\alias{get_genome_enzymes}
\title{Query all enzymes related to each genome in the dataset}
\usage{
get_genome_enzymes(data, pathway_enzymes = NULL)
}
\arguments{
\item{data}{a tibble with genomes and genome_id or the output from `get_genome_id()`}

\item{pathway_enzymes}{a vector of pathway enzymes}
}
\description{
Query all enzymes related to each genome in the dataset
}
\details{
if `pathway_enzymes` is NULL (default) than all enzymes for each genome
will be returned. If a vector of enzyme_ids is provided than only the enzymes related
to each genome that are in this vector will be returned. This is useful if used in
coordination with the output of `get_pathway_enzymes` to only include the enzymes
that are related with the pathway of interest.
}
