% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-genome-id.R
\name{get_genome_id}
\alias{get_genome_id}
\title{retrieve KEGG genome id for all genomes}
\usage{
get_genome_id(data)
}
\arguments{
\item{data}{dataset of genome names to get id for}
}
\description{
retrieve KEGG genome id for all genomes
}
\details{
This took ~4.4 minutes to run for 400 genomes
}
