% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-genome-orthologies.R
\name{get_genome_orthologies}
\alias{get_genome_orthologies}
\title{Query all orthologies related to each genome in the dataset}
\usage{
get_genome_orthologies(data, pathway_orthologies = NULL)
}
\arguments{
\item{data}{a tibble with genomes and genome_id or the output from `get_genome_id()`}

\item{pathway_orthologies}{a vector of pathway orthologies}
}
\description{
Query all orthologies related to each genome in the dataset
}
\details{
if `pathway_orthologies` is NULL (default) than all orthologies for each genome
will be returned. If a vector of orthology_ids is provided than only the orthologies related
to each genome that are in this vector will be returned. This is useful if used in
coordination with the output of `get_pathway_orthologies` to only include the orthologies
that are related with the pathway of interest.
}
