% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-org-id.R
\name{get_org_ids}
\alias{get_org_ids}
\title{retrieve KEGG genome id for all organims}
\usage{
get_org_ids(data, verbose)
}
\arguments{
\item{data}{dataset of genome names to get id for}
}
\description{
retrieve KEGG genome id for all organims
}
\details{
This took ~4.4 minutes to run for 400 organisms
}
