% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-orthologies.R
\name{get_orthologies}
\alias{get_orthologies}
\alias{get_orthologies.keggerator}
\alias{get_orthologies.orgs_id}
\title{Query all orthologies related to each genome in the dataset}
\usage{
get_orthologies(data, pathway_orthologies, kegg_orthology, verbose,
  progress)

\method{get_orthologies}{keggerator}(data, pathway_orthologies = NULL,
  kegg_orthology = NULL, verbose = FALSE, progress = TRUE)

\method{get_orthologies}{orgs_id}(orgs_id, pathway_orthologies = NULL,
  kegg_orthology = NULL, verbose = FALSE, progress = TRUE)
}
\arguments{
\item{data}{a tibble with genomes and genome_id or the output from `get_genome_id()`}

\item{pathway_orthologies}{a vector of pathway orthologies}

\item{kegg_orthology}{a kegg_tbl with the columns orthology and orthology_id
which can be generated with the get_kegg_orthology function. If NULL (default)
the KEGGerator::kegg_orthologies dataset will be used}

\item{verbose}{logical; if TRUE the number of enzyme links for each genome
id will be shown as they are processed}

\item{progress}{logical; if TRUE (default) than a progress bar will appear if
the query takes longer than 10 seconds (it likely will if your data has more)
than only a few genomes}
}
\description{
Query all orthologies related to each genome in the dataset
}
\details{
if `pathway_orthologies` is NULL (default) than all orthologies for each genome
will be returned. If a vector of orthology_ids is provided than only the orthologies related
to each genome that are in this vector will be returned. This is useful if used in
coordination with the output of `get_pathway_orthologies` to only include the orthologies
that are related with the pathway of interest.
}
\section{Methods (by class)}{
\itemize{
\item \code{keggerator}: method for keggerator objects

\item \code{orgs_id}: method for orgs_id objects
}}

