% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-module.R
\name{get_orthology_modules}
\alias{get_orthology_modules}
\title{Query all modules related to each orthology in a dataset}
\usage{
get_orthology_modules(data)
}
\arguments{
\item{data}{a tabble that is output from get_pathway_orthologies()}
}
\description{
Query all modules related to each orthology in a dataset
}
