% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-links.R
\name{get_pathway_enzymes}
\alias{get_pathway_enzymes}
\title{Query all enzymes associated with a given pathway}
\usage{
get_pathway_enzymes(pathway_name)
}
\arguments{
\item{pathway_name}{a character string specifying a certain pathway. Regular
expressions are used to match the pathway so full pathway names do not have
to be given.}
}
\description{
Query all enzymes associated with a given pathway
}
\details{
the value passed to pathway_name is searched against the pathway
column in the kegg_pathways dataset
}
