% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome-funs.R
\name{has_complete_function}
\alias{has_complete_function}
\title{returns genomes that have all the functions in specified module}
\usage{
has_complete_function(data, function_data, complete_functions)
}
\arguments{
\item{data}{a tibble that is output from \code{get_genome_orthologies()}}

\item{function_data}{a tibble containing a function id for each of
the orthologies in a desired pathway. An example of this can be seen in the
\code{nitrogen_metabolism_orth_fun} dataset}

\item{complete_functions}{a tibble containing all of the functions
required for each full process in a given pathway. An example can be seen
in the \code{nitrogen_metabolism_fun_groups} dataset}
}
\description{
returns genomes that have all the functions in specified module
}
\details{
This function takes three tibbles. The first (\code{data}) is a tibble that
is output from \code{get_genome_orthologies}. This contains all of the
orthologies associated with both the genome and the desired pathway.

The second tibble (\code{function_data}) contains the function id and function meaning for each
orthology in a given pathway. This data should be stored in columns named
\code{func_id} and \code{func_meaning}, respectively. There should also be a
column containing the modules that these functions are associated with by
the name of \code{modules_name}.

The third thibble (\code{complete_functions}) should be a tibble with two columns.
The first column should be named \code{modules_name} and there should be
a row for every module in a specific pathway. The second column should be
named \code{funcs_in_module} and should be a list column where each row contains
a list with all of the function ids (corresponding to the \code{func_id}
column in the function_data dataset)
}
