% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome-funs.R
\name{has_specific_function}
\alias{has_specific_function}
\title{find all genomes that complete all of the desired functions}
\usage{
has_specific_function(data, function_data, desired_functions)
}
\arguments{
\item{data}{a tibble containing all of the orthologies for each genome
in a specific pathway. This should be the output of the
\code{get_genome_orthologies} function}

\item{function_data}{a tibble containing a function id for each of
the orthologies in a desired pathway. An example of this can be seen in the
\code{nitrogen_metabolism_orth_fun} dataset}

\item{desired_functions}{a list of vector containing all of the function ids
(corresponding to those in the \code{func_id} column of the function_data
tibble) that a genome should carry out

#' @details This function takes two tibbles and a vector. The first (\code{data}) is a tibble that
is output from \code{get_genome_orthologies}. This contains all of the
orthologies associated with both the genome and the desired pathway.

The second tibble (\code{function_data}) contains the function id and function meaning for each
orthology in a given pathway. This data should be stored in columns named
\code{func_id} and \code{func_meaning}, respectively. There should also be a
column containing the modules that these functions are associated with by
the name of \code{modules_name}.

The third argument (\code{desired_functions}) takes a vector or a list of
function ids (corresponding to thos in the \code{func_id} column in the
\code{function_data} tibble).}
}
\description{
find all genomes that complete all of the desired functions
}
