% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggerator-class.R
\name{keggerator}
\alias{keggerator}
\title{A keggerator object}
\usage{
keggerator(tax_tbl = NULL, otu_tbl = NULL, sam_tbl = NULL,
  otu_ref = NULL, orgs_tbl = NULL, species_uncert = NULL,
  orgs_id = NULL, kegg_uncert = NULL, total_uncert = NULL,
  orgs_filt = NULL, orgs_enzymes = NULL, orgs_orthologies = NULL,
  orgs_genes = NULL)
}
\value{
an object of class keggerator
}
\description{
A class of object that contains all of the
required components for analysis with
KEGGerator
}
