% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{keggerator_uncertainty}
\alias{keggerator_uncertainty}
\alias{keggerator_uncertainty.keggerator}
\alias{keggerator_uncertainty.orgs_list}
\title{Calculate total uncertainty}
\usage{
keggerator_uncertainty(data)

\method{keggerator_uncertainty}{keggerator}(data)

\method{keggerator_uncertainty}{orgs_list}(data)
}
\description{
Calculate total uncertainty
}
\section{Methods (by class)}{
\itemize{
\item \code{keggerator}: method for keggerator objects

\item \code{orgs_list}: method for orgs_list objects
}}

