% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq-to-tibbles.R
\name{otu_ref}
\alias{otu_ref}
\title{Create a otu reference table}
\usage{
otu_ref(data, otu, tax)
}
\arguments{
\item{data}{an object to convert to otu_ref}
}
\description{
Create a otu reference table
}
