% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq-to-tibbles.R
\name{otu_tibble}
\alias{otu_tibble}
\alias{otu_tibble.phyloseq}
\alias{otu_tibble.otu_table}
\title{Convert OTU table to tibble}
\usage{
otu_tibble(data, quiet)

\method{otu_tibble}{phyloseq}(data)

\method{otu_tibble}{otu_table}(data, quiet = FALSE)
}
\arguments{
\item{data}{a phyloseq object}

\item{quiet}{whether to pass warnings about otu matching or not}
}
\description{
Convert OTU table to tibble
}
\section{Methods (by class)}{
\itemize{
\item \code{phyloseq}: method for phyloseq objects

\item \code{otu_table}: method for otu_table objects
}}

