% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble-genomes.R
\name{genomes_tibble}
\alias{genomes_tibble}
\title{extract genomes from tax table or tax tibble}
\usage{
genomes_tibble(data, drop_taxa = TRUE, strict = FALSE)
}
\arguments{
\item{data}{either a phyloseq object or the output of tibble_tax}

\item{drop_taxa}{logical; should the taxonomy be removed from the output}

\item{strict}{logical; if set to TRUE, it will not keep any taxa that have
not been classified to the species level.}
}
\description{
extract genomes from tax table or tax tibble
}
