% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-module.R
\name{get_enzyme_modules}
\alias{get_enzyme_modules}
\title{Query all modules related to each enzyme in a dataset}
\usage{
get_enzyme_modules(data)
}
\arguments{
\item{data}{a tibble that is output from get_pathway_enzymes()}
}
\description{
Query all modules related to each enzyme in a dataset
}
