% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq-to-tibbles.R
\name{otu_tibble}
\alias{otu_tibble}
\title{Convert OTU table to tibble}
\usage{
otu_tibble(data, quiet)
}
\arguments{
\item{data}{a phyloseq object}

\item{quiet}{whether to pass warnings about otu matching or not}
}
\description{
Convert OTU table to tibble
}
